+++ src/apis/common/queries.ts
@@ -1,1 +1,6 @@
-export const useGetExample = () => {};
+import { Keys, getNotification } from '.';
+import { useSuspenseQuery } from '@suspensive/react-query';
+
+export const useGetNotifications = () => {
+  return useSuspenseQuery(Keys.getNotifications(), getNotification);
+};
