+++ src/app/[lng]/(sub)/notification/components/NotificationSection.tsx
@@ -1,0 +1,56 @@
+'use client';
+import NotiCard from './NotiCard.client';
+import { useGetNotifications } from '@/apis/common';
+import { useTranslation } from '@/app/i18n/client';
+import { Icon } from '@/components/Icon';
+import { Flex } from '@/components/Layout';
+import { ItemList } from '@/components/List';
+import { ComponentProps } from 'react';
+
+const DUMMY_DATA: Array<ComponentProps<typeof NotiCard>['notiData']> = [
+  {
+    title: '모임 참여 인원으로 승인되었습니다.',
+    imageUrl: '/images/dummy_image.png',
+    date: new Date().toISOString(),
+    isRead: false,
+  },
+  {
+    title: '1시간 뒤 모임이 진행됩니다. 준비는 되었나요?',
+    imageUrl: '/images/dummy_image.png',
+    date: new Date(Date.now() - 6000000).toISOString(),
+    isRead: true,
+  },
+  {
+    title: '모임이 취소되었습니다.',
+    imageUrl: '/images/dummy_image.png',
+    date: new Date(Date.now() - 86400000 * 2).toISOString(),
+    isRead: false,
+  },
+  {
+    title: '신규 인원이 회원 님의 모임에 지원하였습니다.',
+    imageUrl: '/images/dummy_image.png',
+    date: new Date(Date.now() - 86400000 * 20).toISOString(),
+    isRead: true,
+  },
+];
+
+export default function NotificationSection() {
+  const { i18n } = useTranslation('common');
+  // const { data } = useGetNotifications();
+
+  return (
+    <ItemList
+      data={DUMMY_DATA}
+      renderItem={(data) => <NotiCard notiData={data} />}
+      hasDivider={false}
+      renderEmpty={() => (
+        <Flex direction="column" justify="center" align="center" className="my-80 gap-8">
+          <Icon id="48-cancel" width={48} height={48} />
+          <p className="text-sign-tertiary">
+            {i18n.language === 'ko' ? '알림이 없습니다.' : 'No notifications.'}
+          </p>
+        </Flex>
+      )}
+    />
+  );
+}
